[PATRO23-43] move budgets do shared folder
