+++ src/modules/budget/application/Budget/ExportingBudgets/GetBudgetsToExport.cs
@@ -1,4 +1,3 @@
-using Intive.Patronage2023.Modules.Budget.Application.Budget.ExportingBudgets;
 using Intive.Patronage2023.Modules.Budget.Application.Budget.Mappers;
 using Intive.Patronage2023.Modules.Budget.Infrastructure.Data;
 using Intive.Patronage2023.Modules.User.Contracts.ValueObjects;
@@ -6,7 +5,7 @@ using Intive.Patronage2023.Shared.Abstractions;
 using Intive.Patronage2023.Shared.Abstractions.Queries;
 using Microsoft.EntityFrameworkCore;
 
-namespace Intive.Patronage2023.Modules.Budget.Application.Budget.Shared.ExportingBudgets;
+namespace Intive.Patronage2023.Modules.Budget.Application.Budget.ExportingBudgets;
 
 /// <summary>
 /// Record GetBudgetsToExport representing a query to retrieve budgets ready for export.
