+++ src/modules/budget/application/Budget/ImportingBudgets/BudgetAggregateList.cs
@@ -3,7 +3,7 @@ using Intive.Patronage2023.Modules.Budget.Domain;
 using Intive.Patronage2023.Shared.Abstractions.Commands;
 using Intive.Patronage2023.Shared.Abstractions.Domain;
 
-namespace Intive.Patronage2023.Modules.Budget.Application.Budget.Shared.ImportingBudgets;
+namespace Intive.Patronage2023.Modules.Budget.Application.Budget.ImportingBudgets;
 
 /// <summary>
 /// BudgetAggregateList class represents a command that holds a list of BudgetAggregate.
