+++ src/modules/budget/application/Budget/ImportingBudgets/GetImportResult.cs
@@ -1,6 +1,4 @@
-using Intive.Patronage2023.Modules.Budget.Application.Data;
-
-namespace Intive.Patronage2023.Modules.Budget.Application.Budget.Shared.ImportingBudgets;
+namespace Intive.Patronage2023.Modules.Budget.Application.Budget.ImportingBudgets;
 
 /// <summary>
 /// GetImportResult command.
