+++ src/modules/budget/application/Budget/Shared/Services/CsvBudgetService.cs
@@ -1,14 +1,14 @@
 using CsvHelper;
 using Intive.Patronage2023.Modules.Budget.Application.Budget.ExportingBudgets;
 
-namespace Intive.Patronage2023.Modules.Budget.Application.Data.Service;
+namespace Intive.Patronage2023.Modules.Budget.Application.Budget.Shared.Services;
 
 /// <summary>
 /// CsvService class implements the ICsvService interface and provides methods
 /// for handling CSV data operations such as writing budget information to a CSV file
 /// and generating a local file path for a new CSV file.
 /// </summary>
-public class CsvService : ICsvService
+public class CsvBudgetService : ICsvBudgetService
 {
 	/// <summary>
 	/// Writes a list of budgets to a CSV file at the specified file path.
