+++ src/shared/infrastructure/Intive.Patronage2023.Shared.Infrastructure.csproj
@@ -6,6 +6,8 @@
         <Nullable>enable</Nullable>
     </PropertyGroup>
 	<ItemGroup>
+	  <PackageReference Include="Azure.Storage.Blobs" Version="12.16.0" />
+	  <PackageReference Include="CsvHelper" Version="30.0.1" />
 	  <PackageReference Include="MediatR" Version="12.0.1" />
 	  <PackageReference Include="StyleCop.Analyzers" Version="1.2.0-beta.435">
 	    <PrivateAssets>all</PrivateAssets>
@@ -15,9 +17,6 @@
 	<ItemGroup>
 	  <ProjectReference Include="..\abstractions\Intive.Patronage2023.Shared.Abstractions.csproj" />
 	  <ProjectReference Include="..\domain\Intive.Patronage2023.Shared.Domain.csproj" />
-	</ItemGroup>
-	<ItemGroup>
-	  <Folder Include="Services\" />
 	</ItemGroup>
 
 </Project>
