+++ src/shared/infrastructure/Services/IBlobStorageService.cs
@@ -1,9 +1,7 @@
-using Azure.Storage.Blobs.Models;
 using Azure.Storage.Blobs;
+using Azure.Storage.Blobs.Models;
 
-using Intive.Patronage2023.Modules.Budget.Application.Budget.ExportingBudgets;
-
-namespace Intive.Patronage2023.Modules.Budget.Application.Data.Service;
+namespace Intive.Patronage2023.Shared.Infrastructure.Services;
 
 /// <summary>
 /// IBudgetExportService interface defines a contract for services that handle the exportation of budget data.
